normalize pubsub
