+++ event/event_type.go
@@ -7,7 +7,7 @@ import (
 	"github.com/taubyte/go-sdk/common"
 	http "github.com/taubyte/go-sdk/http/event"
 	p2p "github.com/taubyte/go-sdk/p2p/event"
-	"github.com/taubyte/go-sdk/pubsub"
+	pubsub "github.com/taubyte/go-sdk/pubsub/event"
 )
 
 func (e Event) Type() common.EventType {
@@ -25,12 +25,12 @@ func (e Event) HTTP() (http.Event, error) {
 	return http.Event(e), nil
 }
 
-func (e Event) PubSub() (pubsub.PubSubEvent, error) {
+func (e Event) PubSub() (pubsub.Event, error) {
 	if e.Type() != common.EventTypePubsub {
 		return 0, errors.New("Not a pubsub event")
 	}
 
-	return pubsub.PubSubEvent(e), nil
+	return pubsub.Event(e), nil
 }
 
 // Gives a p2pEvent if the current function is called through p2p
