+++ pubsub/event/types.go
@@ -1,0 +1,4 @@
+package event
+
+type Event uint32
+type EventMessage uint32
