+++ pubsub/node/channel_test.go
@@ -1,4 +1,4 @@
-package pubsub
+package node
 
 import (
 	"testing"
