+++ pubsub/node/publish.go
@@ -1,4 +1,4 @@
-package pubsub
+package node
 
 import (
 	"fmt"
