+++ pubsub/node/publish_test.go
@@ -1,4 +1,4 @@
-package pubsub
+package node
 
 import (
 	"testing"
@@ -6,12 +6,12 @@ import (
 	symbols "github.com/taubyte/go-sdk-symbols/pubsub"
 )
 
-func TestSubscribe(t *testing.T) {
+func TestPublish(t *testing.T) {
 	symbols.MockData{}.Mock()
 
 	channel := &ChannelObject{name: "someChannel"}
 
-	err := channel.Subscribe()
+	err := channel.Publish([]byte("Hello, world!"))
 	if err == nil {
 		t.Error("Expected error")
 		return
