+++ pubsub/node/subscribe_test.go
@@ -1,4 +1,4 @@
-package pubsub
+package node
 
 import (
 	"testing"
@@ -6,12 +6,12 @@ import (
 	symbols "github.com/taubyte/go-sdk-symbols/pubsub"
 )
 
-func TestPublish(t *testing.T) {
+func TestSubscribe(t *testing.T) {
 	symbols.MockData{}.Mock()
 
 	channel := &ChannelObject{name: "someChannel"}
 
-	err := channel.Publish([]byte("Hello, world!"))
+	err := channel.Subscribe()
 	if err == nil {
 		t.Error("Expected error")
 		return
